refactor(Game): create GameRound data structure
move game round validating methods to GameRound class and rename ValidateUtils to CarNames
